+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Allowed headings:
 
 ## [Unreleased]
 
+### Added
+
+* `WarningView` has new attribute `noPadding` which removes the default padding from the component when set to true. When not set, component has `8dp` padding.
+
 ### Changed
 
 * Updated padding on `HeadlineCardView` so that if there are no children added then the bottom padding is the same as a normal card (16dp). If children are added, this padding is increased back to 24dp automatically.